feat!: consolidate Qwik adapter into main package
Expose the experimental adapter from @vp-tw/likftc/qwik and retire the standalone workspace package. Keep Qwik's beta-only compiler workaround isolated while preserving strict checks for stable adapters.

Also remove NPM_TOKEN from the release job so npm trusted publishing can use OIDC.
