+++ .changeset/README.md
@@ -1,5 +1,5 @@
 # Changesets
 
-Every user-visible package change requires a changeset. The stable `@vp-tw/likftc` package and experimental `@vp-tw/likftc-qwik` package release in one fixed version group so their compatible versions cannot drift.
+Every user-visible package change requires a changeset. Stable and experimental framework adapters ship from `@vp-tw/likftc` subpath exports.
 
 Do not run `changeset publish` from a developer machine. Release publishing uses reviewed CI trusted publishing after package and provenance gates pass.
