+++ .changeset/config.json
@@ -5,7 +5,7 @@
   "bumpVersionsWithWorkspaceProtocolOnly": true,
   "changelog": "@changesets/cli/changelog",
   "commit": false,
-  "fixed": [["@vp-tw/likftc", "@vp-tw/likftc-qwik"]],
+  "fixed": [],
   "ignore": [],
   "linked": [],
   "privatePackages": {
