+++ .changeset/merge-qwik-adapter.md
@@ -1,0 +1,5 @@
+---
+"@vp-tw/likftc": patch
+---
+
+Move the experimental Qwik adapter into the `@vp-tw/likftc/qwik` subpath export.
