+++ .changeset/quiet-trees-move.md
@@ -1,6 +1,5 @@
 ---
 "@vp-tw/likftc": patch
-"@vp-tw/likftc-qwik": patch
 ---
 
 Consolidate the stable framework adapters into `@vp-tw/likftc` subpath exports and publish Qwik separately as an experimental adapter.
