+++ COMPATIBILITY.md
@@ -35,7 +35,7 @@ Node 24.14.0 was sufficient for the first isolated Astro spike but is not a vali
 
 Lit 3.3.3 is a required consumer example for `@vp-tw/likftc/web`, not a runtime peer dependency. The native Web Components export must remain usable without Lit.
 
-Qwik support is experimental and optimizer-only. Stable Qwik 1.20.0 excludes Vite 8, while Qwik 2 beta supports it. Qwik 2 beta.36 and beta.37 still fail semantic checking of their SVG JSX declarations under TypeScript 6 and TypeScript 7, and their runtime ESM reads optimizer globals during direct Node evaluation. Only `packages/qwik/tsconfig.json` enables `skipLibCheck`; adapter source, tests, generated declarations, optimizer output, and real-browser identity behavior remain checked. The peer is exact until Qwik 2 becomes stable and both upstream exceptions are removed.
+Qwik support is experimental and optimizer-only. Stable Qwik 1.20.0 excludes Vite 8, while Qwik 2 beta supports it. Qwik 2 beta.36 and beta.37 still fail semantic checking of their SVG JSX declarations under TypeScript 6 and TypeScript 7, and their runtime ESM reads optimizer globals during direct Node evaluation. `packages/likftc/tsconfig.qwik.json` confines `skipLibCheck` to Qwik-specific type checking, while `tsconfig.build.json` applies the same upstream workaround during declaration generation. Stable adapter checks keep `skipLibCheck: false`; adapter source, tests, generated declarations, optimizer output, and real-browser identity behavior remain checked. The optional peer is exact until Qwik 2 becomes stable and both upstream exceptions are removed.
 
 Vue SFC checking uses `vue-tsc` 3.3.7 with an application-local TypeScript 5.9.3 API because that checker cannot load the native TypeScript 7 shim. The same demo's `.ts` source still passes the workspace TypeScript 6 and TypeScript 7 checks, and the SFC checker inherits the strict application config. Re-evaluate this boundary when `vue-tsc` supports the native TypeScript 7 API.
