+++ apps/demos/package.json
@@ -30,7 +30,6 @@
     "@tanstack/react-store": "0.11.0",
     "@tanstack/store": "0.11.0",
     "@vp-tw/likftc": "workspace:*",
-    "@vp-tw/likftc-qwik": "workspace:*",
     "@xstate/react": "6.1.0",
     "alien-signals": "3.2.1",
     "jotai": "2.20.1",
