+++ packages/likftc/tsconfig.browser.qwik.json
@@ -5,5 +5,5 @@
     "jsxImportSource": "@qwik.dev/core",
     "lib": ["ES2023", "DOM", "DOM.Iterable"]
   },
-  "include": ["src/index.browser.test.tsx", "../../test/browser-conformance.ts"]
+  "include": ["src/qwik.browser.test.tsx", "../../test/browser-conformance.ts"]
 }
