+++ packages/likftc/tsconfig.qwik.json
@@ -7,5 +7,5 @@
     "skipLibCheck": true,
     "tsBuildInfoFile": "dist/.tsbuildinfo"
   },
-  "include": ["src/index.ts"]
+  "include": ["src/index.ts", "src/qwik.ts"]
 }
