+++ packages/likftc/tsconfig.qwik.test.json
@@ -4,5 +4,5 @@
     "lib": ["ES2023", "DOM", "DOM.Iterable"],
     "rootDir": "src"
   },
-  "include": ["src/index.ts", "src/index.test-d.ts"]
+  "include": ["src/index.ts", "src/qwik.ts", "src/qwik.test-d.ts"]
 }
