--- packages/qwik/CHANGELOG.md
@@ -1,9 +1,0 @@
-# @vp-tw/likftc-qwik
-
-## 1.0.0-alpha.1
-
-### Patch Changes
-
-- baf3695: Consolidate the stable framework adapters into `@vp-tw/likftc` subpath exports and publish Qwik separately as an experimental adapter.
-- Updated dependencies [baf3695]
-  - @vp-tw/likftc@1.0.0-alpha.1
