+++ scripts/validate-packages.mjs
@@ -5,10 +5,10 @@ import { dirname, join } from "node:path";
 import { fileURLToPath, pathToFileURL } from "node:url";
 
 const root = dirname(dirname(fileURLToPath(import.meta.url)));
-const packageNames = ["likftc", "qwik"];
-const optimizerOnlyPackageNames = new Set(["qwik"]);
+const packageNames = ["likftc"];
+const optimizerOnlyExports = new Set(["./qwik"]);
 const executableExtension = process.platform === "win32" ? ".cmd" : "";
-const budgets = { likftc: 16_000, qwik: 4_000 };
+const budgets = { likftc: 16_000 };
 const tarballDirectory = join(root, ".artifacts", "packages");
 
 function run(binary, arguments_) {
@@ -40,24 +40,32 @@ for (const packageName of packageNames) {
   run("publint", [packageDirectory, "--strict"]);
   run("attw", ["--pack", packageDirectory, "--profile", "esm-only"]);
 
-  if (optimizerOnlyPackageNames.has(packageName)) {
-    assert.match(
-      manifest.peerDependencies?.["@qwik.dev/core"] ?? "",
-      /^2\.0\.0-beta\.\d+$/,
-      `${manifest.name} must pin an exact Qwik 2 beta peer`,
-    );
-    console.log(`Skipping direct Node import for optimizer-only ${manifest.name}.`);
-  } else {
-    for (const [exportName, conditions] of Object.entries(manifest.exports)) {
-      if (exportName === "./package.json") continue;
-      const importPath = typeof conditions === "string" ? conditions : conditions.import;
-      assert.equal(
-        typeof importPath,
-        "string",
-        `${manifest.name} ${exportName} needs an import path`,
+  assert.match(
+    manifest.peerDependencies?.["@qwik.dev/core"] ?? "",
+    /^2\.0\.0-beta\.\d+$/,
+    `${manifest.name} must pin an exact Qwik 2 beta peer`,
+  );
+  assert.equal(
+    manifest.peerDependenciesMeta?.["@qwik.dev/core"]?.optional,
+    true,
+    `${manifest.name} must keep the Qwik peer optional`,
+  );
+
+  for (const [exportName, conditions] of Object.entries(manifest.exports)) {
+    if (exportName === "./package.json") continue;
+    if (optimizerOnlyExports.has(exportName)) {
+      console.log(
+        `Skipping direct Node import for optimizer-only ${manifest.name}${exportName.slice(1)}.`,
       );
-      await import(pathToFileURL(join(packageDirectory, importPath)).href);
+      continue;
     }
+    const importPath = typeof conditions === "string" ? conditions : conditions.import;
+    assert.equal(
+      typeof importPath,
+      "string",
+      `${manifest.name} ${exportName} needs an import path`,
+    );
+    await import(pathToFileURL(join(packageDirectory, importPath)).href);
   }
 
   const packed = pack(packageDirectory);
