+++ tsconfig.json
@@ -1,4 +1,4 @@
 {
   "files": [],
-  "references": [{ "path": "./packages/likftc" }, { "path": "./packages/qwik" }]
+  "references": [{ "path": "./packages/likftc" }]
 }
